core: add support for PKCS#12 certificates and keys (bgo #534219)
Allow PKCS#12 files to be used.  openvpn requires that all of
CA, client cert, and private key be in the same PKCS12 file, so the
UI widgets make sure that either all cert file chooser buttons
contain the selected PKCS#12 file, or none of them do.

This should also clean up handling of PEM private keys so that if
the key is unencrypted, the user is not asked for a password.

(heavily cleaned up and refactored by dcbw)