add support for --float option
Essentially, --float tells OpenVPN to accept authenticated packets from
any address, not only the address which was specified in the --remote
option. This allows remote peer to change its IP address and/or port
number. This is useful when you are connecting to a peer which holds
a dynamic address such as a dial-in user or DHCP client.

[thaller@redhat.com: rebased on current master]
[thaller@redhat.com: add import/export and tests]
[jklimes@redhat.com: fix shortcut clash]

https://mail.gnome.org/archives/networkmanager-list/2010-November/msg00014.html

https://bugzilla.gnome.org/show_bug.cgi?id=737108