properties: preserve secret flags and consolidate secrets handling
Secrets flags weren't being preserved when the connection was edited,
plus they were being ignored if there was no password (ie if the
password was not saved).  This rendered the initial secrets flags from
import operations useless.

When the VPN plugin was called to update the connection with all the
new settings, it would replace the existing VPN setting in the
connection.  Thus all existing secrets flags would get blown away
and all secrets would revert to being system owned.

Fix that by keeping track of secrets flags, and at the same time
simplify setting UI widgets with secrets, and pulling the changed
secrets out of the UI widgets for addition to the VPN setting when
updating the connection.