properties: remove dead code and simplify password widget setup
Since the secret agent (ie nm-applet) takes care of getting secrets
for us (via the auth dialog), saving updated secrets, and deleting
old ones, the VPN plugin doesn't need to do that itself anymore.
So we get to delete a bunch of code that extracted and/or deleted
secrets from the keyring since the agent handles all of that.  Only
the auth dialog needs to talk to gnome-keyring anymore.