auth-dialog: always return found secrets
Always return any secrets which were found even if not strictly
required.  For ExternalUiMode, if the secret doesn't need to be
requested from the user, ensure that ShouldAsk is FALSE for
secrets which are found, but for which the user doesn't need
to enter, such as saved secrets.

This fixes a problem with GNOME Shell when multiple secrets are
stored by the user agent.  Since agents delete any saved secrets before
saving new secrets, this could delete some secrets, but since
that item was not included in the returned secrets, it wouldn't be
re-saved, and would be lost.