auth-helpers: correct the selection of certificates
The commit 522190ad916b301919535208193b9d69db0f5f8e fixed an endless loop. But
it didn't make the PKCS#12 vs. non-PKCS#12 file selection right. This commit
makes sure that either all or none of the certificates are in PKCS#12 format,
as it is required.

Note:
The bug mentioned in the commit 522190ad (rh #997255) might not be the cause
of the endless loop problem. But rather it is a defect in Gtk:
https://bugzilla.redhat.com/show_bug.cgi?id=1091837

(cherry picked from commit 62bf1c8875908d8550238009775eb31de4e4ebdc)