core: track openvpn state to detect reconnections
Unfortunately openvpn doesn't allow us to suppress pushed
'ping-exit' options, in which case the exit action will
override our default 'ping-restart' and openvpn will just
exit instead of restarting as we really want.

But, if the server is correctly configured, this commit
will cause the plugin to re-enter the ACTIVATING state,
allow a reconnect, and update NetworkManager with the new
IP configuration, all while leaving the tunnel up.