core: use a default renegotiation interval of zero (rh #969433)
Since the client and server do not negotiate options, each side gets
to specify its own --reneg-sec to control when each side renegotiates.
OpenVPN defaults to 3600, so if the client and server don't agree this
causes too-frequent renegotiations.

This is worse with two-factor authentication, becuase it can mean that
the client requests a password/PIN from the user much more often then
the server actually wants.

https://bugzilla.redhat.com/show_bug.cgi?id=969433