properties: do not require password for always-ask, not-required (bgo #756638)
If password is marked as always-ask or not-required, do not insist on having a
password.
Actually we might detect password requirement incorrectly, because we regard
all PKCS#12 private keys as encrypted. Thus we require a password even if
the private key is not really encrypted. This commit allows user to override
the bad detection.

https://bugzilla.gnome.org/show_bug.cgi?id=756638