properties: remove g_warning() during do_export()
Printing a warning is not a good way to report an error to the
user from the plugin. Either come up with a proper warning/error-reporting
mechanism, or just silently proceed.

Indeed, we have not strong verification for VPN connections, thus
invalid connections are to be expected and the plugin should handle
them best-effort without nagging.