properties: write inline certs only at the end of successful import
Don't write the parsed inline certificates to file right away.
Instead remember them and only write them out at the very end
when the parsing proved to be successful. Otherwise, the import
might still fail afterwards when we already wrote the files.

do_import() writing files to disc is still very ugly, as it is
intransparent to the user, possibly writes to the wrong location
(overwriting wrong file) and the user cannot configure it (whether he
wants a different location or not persistating at all).
Also, once written certifications will not be automatically
purged after the connection will be deleted.

Eventually, the import API should become more powerful so that
inline blobs and filenames can be propagated back and the caller
can decide what to do with them (e.g. by asking the user).