service: fix handling child processes and their cleanup
Previously, ensure_killed() would never be canceled, meaning
2 seconds after sending SIGTERM to the openvpn process, we would
send a SIGKILL signal (to whatever process happens to be around).

Also, the child-watch callback would sooner or later be invoked
when the process finally dies. But since we never unsubscribe
the watcher, we can get the signal for a previously died process.
Also, calling waitpid() is wrong inside a child-watch since the process
was already reaped.

Rework this and keep track of the pending processes in a list
@pids_pending_list. The currently active process is the one
with matching priv->pid.

On shutdown, we wait for all processes to terminate (and still sent
them the final SIGKILL as necessary).