import: ignore file encoding for ovpn configuration file
Openvpn treats ovpn files as ASCII configuration file and
does not care about a specific certain encoding. As such,
only encodings that are an extension of ASCII can work at
all (like iso8859-* or utf8).

We should not try to handle configuration files that cannot even
be handled by openvpn itself.

As regular options must be ASCII-compatbile, the encoding only
matters for options, filenames and inline-blobs.

Inline blobs usually are ASCII/base64 encoded. If they happen to be in a
different encoding, we still want to preserve the original blob and
not guess and convert encodings.

Openvpn itself doesn't care about encoding of filenames and passes
them directly to the system functions (open, access). The same is true
for glib, which expects paths in "GLib file encoding".
Note, that openvpn doesn't support escape sequences like "\344", thus,
if the user really wanted to specify such a character, he is only
able to do so if we don't mess with the encoding during import.

Filenames are now handled as "utf8safe" properties, so we can keep
them in their original encoding.

The only sane option is ignoring the encoding and pretend it is
ASCII compatible. Who writes non-utf8 configuration files anyway?