properties/tests: use separate output directory for tests
When building in-tree, SRCDIR and TMPDIR would point to the
same directory. Which means that test_tls_inline_import()
would overwrite the configuration files to which it tries
to assert against.

For out-of-tree builds, the directories differed. Commit
b2c41184de73c tried to fix a failure of `make distcheck`.
That however is wrong for the in-tree case, where it would
delete files from the source tree.