all: check for UTF-8 encoding and escape invalid characters in filenames
NMSettingVpn is a dictionary of strings which gets serialized over
D-Bus. Thus all values must be UTF-8.

Ensure that during import we only set UTF-8 values.

Certain properties, like the device name and filenames are not in a specific
encoding. Rather they are NUL terminated C-string that are directly
passed to the encoding-agnostic C/kernel API (like open() or the interface
name).
These properties are not plain UTF-8 strings but must be escaped using
the nmv_utils_str_utf8safe_*() functions. With this, unless a string
contains a non-UTF-8 character or the escaping character '\\', they
look just like regular UTF-8 values.

From a UI point of view, the user can now edit arbitrary names
in the input field by using backslash escapes. The GTK file chooser
should be adjusted to properly convert back and forth. This is not yet
done.

The following properties are utf8safe escaped:

  - NM_OPENVPN_KEY_DEV
  - NM_OPENVPN_KEY_CA
  - NM_OPENVPN_KEY_CERT
  - NM_OPENVPN_KEY_KEY
  - NM_OPENVPN_KEY_STATIC_KEY
  - NM_OPENVPN_KEY_TA