shared: rename "nm-openvpn-service-defines.h" to "nm-service-defines.h"
For nm-vpnc (and in the future other VPN plugins), we have a similar
file that serves the same purpose.

Instead of naming them different for each VPN plugin, give them all the
same name (which makes it more obvious how they relate, and it's more
convenient when looking at a diff between plugins).