build: move copied shared files to separate directory
Some of the shared files are copied as-is from NetworkManager's
repository or between VPN plugins. Those files are special, in that
they are not supposed to be edited locally, without considering that
they are used in other projects.

Move them to a separate sub-directory, to make the nature of these
files clearer.