service/trivial: rename logging macro _LOGd() to _LOGD()
Currently, the _LOGD() is differnt from _LOGI() and _LOGW(),
in that it logs conditionally based on the debug setting.
The other two logging macros always log their message.

Later these macros will be unified, so that also _LOGI and _LOGW
considers the logging level. So rename them now to make the later
diff smaller.

Note that in NetworkManager core project, _LOGt() is a statement
that can be disabled at compile time. Thus, _LOGd() had a different
meaning here.