all: use libsecret to wipe the password memory
The passwords can still be paged away to secondary storage. libgcr
provides API to lock the ranges into memory; however it has a dependency
chain that's too big for the service (it drags in Gtk).

Nevertheless, the memset() was completely useless; it would just be optimized
away at compiler's liberty.