connection: add support for verify-x509-name property
The verify-x509-name option deprecates the tls-remote option starting
with OpenVPN 2.3. Add support for it in the connection properties, as it
is an important security feature for avoiding man-in-the-middle attacks.

The connection property is stored as `{type}:{value}`, with type being
one of (name, name-prefix, subject), and value a comma-separated
sequence of `key=value` comparisons. It is combined/split-up into the
two OpenVPN option parameters when import/exporting.