properties: add support for verify-x509-name
Expose the verify-x590-name option in the editor UI by reusing the
tls-remote subject entry, and adding a combo box for selecting a
verification mode.

Selectable modes map to their respective verify-x509-name types
('subject', 'name' or 'name-prefix'), except for 'legacy', which sets
tls-remote and omits verify-x509-name, and 'none', which omits both.

Descriptions and tooltips are updated to attempt to describe the
different modes in a straightforward way, but their actual operation is
unfortunately not very simple.

Updates to translations will be necessary later.