properties: ignore case when comparing cipher and hmac options
openvpn accepts both the lower and uppercase of many ciphers and hmac
algorithms, for example:

 AES-128-CBC, aes-128-cbc
 SHA1, sha1
 ...

Other values instead must be in the right case as:

 ecdsa-with-SHA1
 whirlpool

Let's use a case-insensitive compare of the values. Note that this
also matches values as 'AES-128-cbc' or 'Whirlpool' which are not
accepted by openvpn, but the only alternative would be to hardcode all
possible values in the plugin, which we don't want to do.