properties: add GUI support for comp-lzo "no" and "adaptive"
openvpn and the rest of the plugin already supports all for possible
options for the comp-lzo setting:
  0) option unspecified
  1) "yes"
  2) "adaptive"
  3) "no"

The GUI however only supported a boolean checkbox, which mapped to
the choices 0) (option unspecified) and 1) ("yes"). Obviously, it was
unable to express the other two choices.

Especially problematic is that the GUI previously would wrongly coerce
option 3) ("no") to option 0) (unspecified). That is wrong because it
breaks the setup when the server enables compression. A more correct way
for the GUI would be to map all unsupported options to "yes".

Instead of doing that, just expose all supported comp-lzo options in the
GUI.

https://bugzilla.gnome.org/show_bug.cgi?id=778123