build: adjust required libnm version to 1.7.0
Currently this is nm-openvpn-1.7.0 development version. It requires at least
the (unstable) libnm version 1.7.0. It should not be prevented to build the
plugin against libnm-1.7.0.

Obviously, nm-openvpn-1.7.0 requires libnm-1-8 API. However, the
development version libnm-1-7 may or may not work. If the user decides
to build against a development version, it's up to him to ensure that
the unstable libnm version has the right API. But nm-openvpn should
not prevent that.

Arguably, at this point it doesn't matter much because libnm-1.8.0 is
released for a long time. Still fix it so that on the next version bump
it's clear that the plugin should allow building against the development
version.