properties: cleanup compatibility wrapper for nm_utils_is_valid_iface_name()
libnm deprecated nm_utils_iface_valid_name() for nm_utils_is_valid_iface_name().
In hindsight, that was a mistake because there wasn't really anything wrong
with the existing function which requires to discourage it's use.

Anyway, so to build the same code against libnm and libnm-glib, we need to
add a compatibility wrapper *sigh*.

It is wrong that the nm_utils_is_valid_iface_name() compatiblity macro has
an output argument @e, which is silently dropped. It means, when building
against libnm-glib, the compatiblity wrapper behaves differently. That
would lead to a crash when somebody tried to make use of the error
output argument.

Instead, the macro shall have no error argument. At that point, it shall
also not have the same name as nm_utils_is_valid_iface_name(), to avoid
confusion.