properties: merge "properties/auth-helpers.c" in "nm-editor-plugin.c"
In the past, the properties plugin contained the non-gtk part and the
GTK gui. Hence, several GTK related parts were in "properties/auth-helpers.c".

That is no longer the case. The GTK part of the plugin is in
"properties/nm-openvpn-editor.c".

There is no reason for this split. Moving related code to different
files doesn't make it simpler but more complicated.

Move the GTK code back to "properties/nm-openvpn-editor.c"

No code was changed, except necessary adjustments.