properties: don't require username in GUI
The connection is perfectly valid without the "username" key.
Note that the VPN settings support a data setting
NM_OPENVPN_KEY_USERNAME ("username"), but NMSettingVpn also has a
NM_SETTING_VPN_USER_NAME ("user-name") setting.

So, NM_OPENVPN_KEY_USERNAME has preference over
NM_SETTING_VPN_USER_NAME. But even if the username is entirely missing
in the setting, NetworkManager will populate it with the user-name
of the user activating the connection.

It is wrong to require a user name in the GUI.