Revert "service: keep reference to plugin while main loop is running"
After leaving the main loop on SIGTERM, we want to tear down the
NMVpnServicePlugin instance. This should also unexports the service on
D-Bus and ensure that no more requests are accepted -- as we wouldn't
handle them anyway.

Note that afterwards we still want to iterate the main loop waiting for
processes to terminate.

This reverts commit 9f45977058f7c8cf1fb0c67636cab43e2c85f228.
The crash that 9f45977058 tried to address is actually a bug
in libnm, where NMVpnServicePlugin would not unexport the D-Bus
service. This must be fixed by libnm instead [1]. Note that overall
the crash is of limited priority, because we are already in the process
of shutting down.

[1] https://cgit.freedesktop.org/NetworkManager/NetworkManager/commit/?id=929f36c56f3b22a917066c5cb7bfc938630f9e8c