all: asciibetically sort property list and defines
Reordering #define has no visible effect.

Reordering the properties like in @advanced_keys has effects in
the order in which we find keys. Optimally we would have frequently
used keys as first. But in reality, if the order of the search list
matters performance we, we should instead use a lookup table. So,
asciibetically sort oder is suitable also here and looks consistent
in source code.