all: add support for --extra-certs options
By looking at the source code, it seems that openvpn accepts extra-certs
for every configuration type. On the other hand, the manual page groups
the option under TLS Mode, indicating that this only makes sense for
TLS. In the GUI and during export, handle extra-certs only for TLS
connection. When starting openvpn, pass it one whenever it's present
in the connection.