properties: improve tooltip for tls-crypt certificate button
The same butten is responsible for both for the tls-auth and tls-crypt option.
Maybe, the tooltip should completley change, depending on the mode selected
above. However, that is more complicated and not necessarily better. Swapping
the tooltip could also be confusing. So, go the easy way, and describe both
options in the same tooltip.