properties: move nmovpn_arg_is_set() helper to shared
nmovpn_arg_is_set() expresses the concept of whether a data key is present,
whether it should be considered as set or unset. That is, in many cases an
empty value is treated like a missing/unset key.

This concept is not only revevant during export, but everytime we access
the setting, like when initializing the GUI or when passing the argument
to openvpn.

Hence, move it to shared/utils.h.