service: use nmovpn_arg_is_set() to check whether key is set
We should be consistend about when we consider a key present and when
it is missing. Incidentally, for certain properties we treat an empty
value like it is missing/unset.

This check seems arbitrary and is done inconsistently. At least,
use a particular function nmovpn_arg_is_set() which embodies the
requirement and which is also used by export() in the same context.