properties: sanitize newlines ('\n') when exporting ovpn file
Openvpn's ovpn configuration files cannot contain newlines as values,
because

- the parser reads the ovpn file line by line, splitting lines are
  newlines. Hence, quotation spaning multiple lines is not supported
  to express '\n'.
- special escape sequences like "\n" or "\012" are not supported
  either. Escape sequences are only for "\'", "\"", and verbatim
  spaces like "\ " (but not newline itself).

Note that newlines can be part of a valid configuration, for example
as part of filenames.

Instead of breaking the line, still try to escape the newline
with "\n".

Previously, openvpn would fail with

    $ openvpn --config my.ovpn
    Options error: No closing single quotation (') in my.ovpn:1

Now, the configuration file has no more such line breaks and openvpn
issues a warning instead:

    $ openvpn --config my.ovpn
    Options warning: Bad backslash ('\') usage in my.ovpn:1: remember that backslashes are treated as shell-escapes and if you need to pass backslash characters as part of a Windows filename, you should use double backslashes such as "c:\\openvpn\\static.key"

The point of this change is not that the new warning behavior is better
than the previous error behavior (it's arguably worse). However, I think
if we write an invalid file, it's better to have an invalid escape sequence,
instead of a spurious line break.

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/merge_requests/5