service: pass "--auth" parameter directly to openvpn without validation
The "auth" parameter is the digest that openvpn directly passes
to the crypto library (openssl or mbedtls). It really depends
on what the crypto library supports. For example, openssl
is fine with both "SHA1" and lower case "sha1".

Don't do any validation of the parameter, just pass it to openvpn.
Before, a lower case value like "sha1" would have been rejected.
There is no need for the plugin to understand and validate the
selected auth parameter.

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/issues/8
https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/merge_requests/7