editor: when editing show unknown "auth" parameter in combo box
The "auth" parameter really depends on the SSL library which
openvpn uses. We should accept values gracefully. In the GUI,
the parameter is a combobox, so we must know which values
are valid there. However, when editing an existing profile (for
example, imported from .ovpn file), also accept an unknown digest
and list it as valid selection in the GUI.

This allows a user to set "auth" names which are unsupported by the
GUI, and later editing the connection will not destroy the value.

Also, get rid of HMACAUTH_COL_DEFAULT. It's not needed, because we
can encode the default by having a %NULL HMACAUTH_COL_VALUE.

Also, fix leaking "hmacauth" string in advanced_dialog_new_hash_from_dialog().

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/issues/8
https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/merge_requests/7