properties: drop file extension check when importing a file
We don't enforce any policy on the file name/extension when
exporting a connection, but we do it when importing one.
So, one could end up exporting a connection succesfully and
not being able to import it back.
Example:
  ~]$ nmcli connection export ovpn_test pippo.vpn
  ~]$ nmcli connection import file pippo.vpn  type openvpn
  Error: failed to import 'pippo.vpn': unknown OpenVPN file extension.

Moreover, when importing, we enforce the user to specify the connection
type (openvpn in this case).

Drop the policy on the file name extension when importing openvpn
connections.