properties: split steps for filtering and writing blobs in do_import()
Currently, during import we generate the list of blobs we want to
write to disk. At the end, we write them.

That is not great. For example, with `nm-connection-editor --import $FILE`
we already write the blobs to disk very early. The user might still abort
adding the connection, but we already wrote them. Also, if we fail to add
the connection later on, we don't clean up the files.

Instead, we should extend the API and return the blobs to the caller, so
that the caller may handle them accordingly.

In a first step, split the filtering of the list of blobs from writing
them to disk.