Handle the various pkcs12/cert/key/ca combinations.
OpenVPN allows supplying PKCS#12 client cert/key *or* PEM client cert/key, as
well as a PEM CA cert in either case. When supplying --pkcs12, the args --cert
and --key can not be used. So, you can have one of the following:

    - PKCS#12 combined cert/key/ca (one file): --pkcs12
    - PKCS#12 combined cert/key/ca with PEM CA: --pkcs12, --ca
    - PEM cert/key, PEM CA: --cert, --key, --ca

If a user tries to import an invalid config, they will be met with the appropriate
error message. The editor dialog also enforces the PKCS#12/PEM restrictions.

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/-/merge_requests/29