Accept all valid 'proto' arguments.
OpenVPN allows for a bunch of different args for 'proto', namely 'udp' and 'tcp'
variants that include '4' or '6', and specific to tcp, '-client' suffixes. This
was already defined in shared/util.h; this commit changes the 'proto' check to
also use the 'NMOVPN_PROTCOL_TYPES' array, and adds test for all supported protos.

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/-/issues/59

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/-/merge_requests/30