all: add support for "proto" flag
Previously, only "proto-tcp" was supported, which could be a boolean
(anything or "yes").

Obsolete the old flag, and add a "proto" key. If it is present,
"proto-tcp" is ignored.

The GUI still only has a boolean checkbox and cannot express all the
options. Editing a profile in the GUI always sets either "udp" or
"tcp-client".

Obviously, it's ugly to deprecate an existing option, because all the
code now needs to support the redundant, old option (we will never
remove deprecated options, so deprecation merely means that it's not
the best way to do it, not that something will break in the future).
It's ugly, but still acceptable.