helper: ignore IPv6 configuration without an address
OpenVPN 2.5.0 started to pass incomplete IPv6 configurations when the
server is not fully configured for IPv6 but has some IPv6
directives. For example, it is enough to add 'push "route-ipv6 ::/0"'
to a IPv4-only setup to trigger this behavior.

As a result, NetworkManager rejects the configuration as invalid and
the connection fails.

Fix this by ignoring IPv6 configurations without an address.

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/-/issues/64