editor: port to Gtk4
Large part of the line toll went into dealing with removal of
GtkFileChooserButton from Gtk4. Without a UX redesign we just
got to reimplement it using GtkButton with and GtkLabel that
pops up a GtkFileChooserDialog.

Other than that it's mostly the usual story of replacing the
deprecated widget and behavior with more modern one, providing
compatibility shims to keep compatibole with Gtk3.