import: fix importing profiles with PKCS#12 CA
'have_ca' indicates whether we have an explicit CA cert; don't set it
when the CA is from a PKCS#12 file.

Without this fix, trying to import a profile with a PKCS#12 that
includes the CA certificate results in error:

 "--ca can not be PKCS#12 format"

because we first set have_pkcs12=TRUE, have_ca=TRUE and then we check
that the CA is not in PKCS#12 format.

Fixes: 34a0c9257aac ("Handle the various pkcs12/cert/key/ca combinations.")

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/-/issues/83
https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/-/merge_requests/50