properties: add `tls-cipher` to list of advanced parameters
This value being missing from the hash means that if it is added to the
*.nmconnection file, despite being valid it will be removed if any
operation triggers the GUI (i.e. like toggling the link active).

This patch adds it to the initial hash table lookup, and includes a
freetext field under the advanced tab to allow it to be entered.

Users who need this field will find details online, and the tooltip
text makes how to set it clear.

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/-/merge_requests/53