helper: handle DOMAIN-SEARCH in addition to DOMAIN
OpenVPN has suppport for (multivalued) DHCP option DOMAIN-SEARCH (119),
which is ignored by nm-opevpn.

Core NetworkManager doesn't have an option for them either.

However, in practical terms it appears to make sense to just append them
to the domains list.

So handle DOMAIN-SEARCH options mostly similar to DOMAIN ones.  Since
it's not unlikely that some servers may use multiple DOMAIN and
DOMAIN-SEARCH values, accumulate the latter ones in a separate array
first and then append it to the tail of the domains array.

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/-/merge_requests/57