service: fix automatically adding "cipher" to "data-ciphers"
First, the "data-ciphers" in openvpn defaults (depending on version
and compile time configuration) to "AES-256-GCM:AES-128-GCM" or
"AES-256-GCM:AES-128-GCM:CHACHA20-POLY1305". So taking "cipher"
and just set it as "--data-ciphers $CIPHER" is wrong.
When doing this compat work, we will now set "--data-ciphers"
to "X:AES-256-GCM:AES-128-GCM". This should work, and if it's not
suitable to the user, they still can set "data-ciphers" explicitly.

Also, when we do the compat works, it's wrong to still set "--cipher"
to whatever it was. Instead, only set "--data-ciphers" for 2.5 and
newer.

Note that:

- if we fail to detect the openvpn version, we assume to run an old
  version and just set "--cipher" directly. That is unlike more cases
  where we fallback to the most recent behavior. In this case, "cipher"
  is a legacy option, and we assume that we run a legacy openvpn, unless
  we detect otherwise.

- 2.6 includes CHACHA20-POLY1305 in the default "--data-ciphers". The
  plugin doesn't do that in the compat path. The reason is not strong,
  but it would mean to differentiate between 2.5 and 2.6+. Also,
  whether CHACHA20-POLY1305 is enabled by default depends on whether
  it's enabled in OpenSSL. For consistency, the compat path for "cipher"
  follows the 2.5 behavior and that's it.

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/-/issues/112

Fixes: 020ab0c4b872 ('service: automatically add the "cipher" to the "data-ciphers"')