service: fix signal cleaning when getting SIGTERM or SIGINT
The main function cleans up the signals handling SIGTERM and SIGINT. The
signal_handler callback should return G_SOURCE_CONTINUE so the source
won't be removed.

This won't have any negative effect because we are already quitting the
loop before continuing in signal_handler callback and it is avoiding the
double source removal.

https://bugzilla.redhat.com/show_bug.cgi?id=2214996

https://gitlab.gnome.org/GNOME/NetworkManager-openvpn/-/merge_requests/69