service: fix challenge-response not being sent to server
In interactive mode, io_data->password will be NULL when the
challenge-response is received. Instead of checking if we have a
password and a challenge, just check if we have the challenge's reponse
and, if we do, send it to the ovpn server.

Also, refactor a bit the part where the hints are filled. Again, we
don't need to check that we don't have a password: if we have a
challenge but not the response, request the challenge-response.