test: remove usage of deprecated nm_utils_is_uuid
This function is deprecated in libnm, so we get a warning when compiling
the test which can make the compilation to fail.

As there is no replacement for that function in libnm, just remove it's
single usage, used to verify that a uuid parameter was valid. Let's just
assume that it is valid.