editor: send back the challenge response with the tag
The nm-openvpn service, when there is a 2FA dynamic challenge enabled,
requests the secret to the daemon with a prefix tag
"x-dynamic-challenge(-echo):". This tag is an indication that the secret
needs some special handling because it's a 2FA token.

When it was introduced in this plugin and in libnm, it was done in a way
that the secret agent (like the plugin's editor) should remove the
prefix tag. See NetworkManager's commit 345bd1b18731 ('libnmc: fix secrets
request on 2nd stage of 2FA authentication') and commit 27c701ebfbc9
('libnmc: allow user input in ECHO mode for 2FA challenges').

That was not a great idea because it's not backwards compatible so all
the secret agents need to adapt. This has been fixed in NetworkManager's
commit 25dd5f8713c4 ('vpn: handle hint tags in the daemon') and commit
87a1dfccfa57 ('libnmc: don't strip prefix tags from secret names'). Now,
the secret agents are expected not to strip the prefix tag, and let the
daemon to do it.

Change the editor to do that.